refactor: move RustyConnector stuff into own class
